Scene
Created scene object.
Fixes in transform logic.
Added temporary components.
Removed some constatns from ScriptEngine conde.

+++ EngineQ/EngineQ/Scripting/API_Transform.cpp
@@ -75,7 +75,7 @@ namespace EngineQ
 
 		void API_Transform::API_GetChildCount(Transform& transform, int& count)
 		{
-			count = transform.GetChildCount();
+			count = static_cast<int>(transform.GetChildCount());
 		}
 
 		void API_Transform::API_GetChild(Transform& transform, int index, MonoObject*& child)
